[FIX]: theme setup
