+++ README.md
@@ -67,6 +67,73 @@ Update `tsconfig.json`
 ```
 
 
+# Install these helper dependencies
+
+##### 1. react-native-gesture-handler
+
+```javascript
+yarn add react-native-gesture-handler
+```
+
+To finalize the installation of `react-native-gesture-handler`:
+ 1. create a directory inside root folder - `gesture-handler`
+ 2. conditionally import `react-native-gesture-handler` by creating two files inside `gesture-handler`
+
+`gh.native.js`
+
+```javascript
+// Only import react-native-gesture-handler on native platforms
+import 'react-native-gesture-handler';
+```
+
+`gh.js`
+
+```javascript
+// Don't import react-native-gesture-handler on web
+```
+
+Add the following line at the top of your entry file, such as `index.js` or `App.js`:
+
+```javascript
+import './gesture-handler/gh';
+```
+
+After installation, wrap your entry point with <GestureHandlerRootView>
+
+```javascript
+import {GestureHandlerRootView} from 'react-native-gesture-handler';
+
+export default function App() {
+  return (
+    <GestureHandlerRootView style={{flex: 1}}>
+      {/* content */}
+    </GestureHandlerRootView>
+  );
+}
+```
+
+##### 2. react-native-reanimated
+
+```javascript
+yarn add react-native-reanimated
+```
+
+Add the `react-native-reanimated/plugin` to your `babel.config.js`:
+
+```javascript
+module.exports = {
+  presets: [
+    // other presets
+  ],
+  plugins: [
+    // other plugins
+    'react-native-reanimated/plugin',
+  ],
+};
+```
+
+
+
 ### 4. Add some important packages
 
 **svg/image/lottie/gif** -> check README.md inside `assets`
