+++ src/views/Home/useHomeScreen.ts
@@ -1,7 +1,8 @@
 
-import {useAppTheme, useAppTranslation} from '@app/hooks';
 import {useMemo} from 'react';
 import homeScreenStyles from './HomeScreen.style';
+import {useAppTheme} from '@app/theme';
+import {useAppTranslation} from '@app/i18n';
 
 const useHomeScreen = () => {
   const theme = useAppTheme();
