+++ src/views/Login/useLoginScreen.ts
@@ -1,5 +1,5 @@
 
-import {useAppTheme} from '@app/theme';
+import {useAppTheme} from '@app/theme/themes';
 import {useMemo} from 'react';
 import loginScreenStyles from './LoginScreen.style';
 import {useAppTranslation} from '@app/i18n/utils';
