+++ src/views/Splash/useSplashScreen.ts
@@ -1,7 +1,7 @@
 
 import {useEffect, useMemo, useState} from 'react';
 import splashScreenStyles from './SplashScreen.style';
-import {useAppTheme} from '@app/theme';
+import {useAppTheme} from '@app/theme/themes';
 import {useAppTranslation} from '@app/i18n/utils';
 
 const SPLASH_TIMEOUT = 5000;
